Added remaining CRUD methods to CommentService and CommentController.

+++ src/main/java/nycto/homeservices/dto/commentDto/CommentUpdateDto.java
@@ -12,6 +12,6 @@ public record CommentUpdateDto(
         @NotNull(message = "rating can't be null!")
         @Min(value = 1, message = "rating must be between 1 and 5!")
         @Max(value = 5, message = "rating must be between 1 and 5!")
-        Integer rating
+        Double rating
 ) {
 }
